rename some deploy files
